+++ util/src/main/java/edu/jhu/hlt/concrete/serialization/Concretable.java
@@ -1,0 +1,8 @@
+package edu.jhu.hlt.concrete.serialization;
+
+import org.apache.thrift.TBase;
+import org.apache.thrift.TFieldIdEnum;
+
+public interface Concretable<T extends TBase<T, ? extends TFieldIdEnum>> {
+  public T toThrift();
+}