+++ validation/src/main/java/edu/jhu/hlt/concrete/validation/ff/entity/ValidEntityMention.java
@@ -1,0 +1,8 @@
+package edu.jhu.hlt.concrete.validation.ff.entity;
+
+import edu.jhu.hlt.concrete.validation.ff.FlattenedTokenRefSequence;
+import edu.jhu.hlt.concrete.validation.ff.UUIDable;
+
+public interface ValidEntityMention extends UUIDable, FlattenedTokenRefSequence {
+
+}