+++ validation/src/main/java/edu/jhu/hlt/concrete/validation/ff/entity/ValidEntityMentionSet.java
@@ -1,0 +1,10 @@
+package edu.jhu.hlt.concrete.validation.ff.entity;
+
+import java.util.List;
+
+import edu.jhu.hlt.concrete.validation.ff.FlattenedMetadata;
+import edu.jhu.hlt.concrete.validation.ff.UUIDable;
+
+public interface ValidEntityMentionSet extends UUIDable, FlattenedMetadata {
+  public List<ValidEntityMention> getEntityList();
+}