+++ validation/src/main/java/edu/jhu/hlt/concrete/validation/ff/situation/ValidSituationSet.java
@@ -1,0 +1,7 @@
+package edu.jhu.hlt.concrete.validation.ff.situation;
+
+import edu.jhu.hlt.concrete.validation.ff.UUIDable;
+
+public interface ValidSituationSet extends UUIDable {
+
+}