+++ validation/src/main/java/edu/jhu/hlt/concrete/validation/ff/structure/NecessarilyUniqueUUIDSentence.java
@@ -1,0 +1,17 @@
+package edu.jhu.hlt.concrete.validation.ff.structure;
+
+import edu.jhu.hlt.concrete.Sentence;
+import edu.jhu.hlt.concrete.validation.ff.AbstractConcreteStructWithNecessarilyUniqueUUIDs;
+import edu.jhu.hlt.concrete.validation.ff.InvalidConcreteStructException;
+
+public class NecessarilyUniqueUUIDSentence extends AbstractConcreteStructWithNecessarilyUniqueUUIDs<Sentence>
+    implements ValidSentence {
+
+  private final ValidTokenization tkz;
+
+  public NecessarilyUniqueUUIDSentence(Sentence s) throws InvalidConcreteStructException {
+    super(s);
+    this.addNecessarilyUniqueUUID(s.getUuid());
+    this.tkz = Tokenizations.validate(s.getTokenization());
+  }
+}