+++ validation/src/main/java/edu/jhu/hlt/concrete/validation/ff/structure/ValidTokenization.java
@@ -1,0 +1,5 @@
+package edu.jhu.hlt.concrete.validation.ff.structure;
+
+public interface ValidTokenization {
+
+}